requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.8
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
Pillow==9.5.0
pycryptodome==3.17
pyparsing==3.0.9
pyTelegramBotAPI==4.11.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.5.5
requests==2.30.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
statsmodels==0.14.0
telebot==0.0.5
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==2.0.2
websockets==11.0.3
yarl==1.9.2
